fix: check if player is online before add to cache

+++ src/main/java/net/smoothplugins/smoothusers/listener/PlayerJoinListener.java
@@ -47,6 +47,8 @@ public class PlayerJoinListener implements Listener {
 
             if (userService.cacheContainsByUUID(user.getUuid())) {
                 Bukkit.getScheduler().runTaskLaterAsynchronously(plugin, () -> {
+                    if (!event.getPlayer().isOnline()) return;
+
                     if (!userService.removeTTLFromCacheByUUID(user.getUuid()) | !userService.removeTTLFromCacheByUsername(user.getLowerCaseUsername())) {
                         userService.loadToCache(user);
                     }
